+++ src/components/UploadDocumentCard/RightContent/index.tsx
@@ -1,0 +1,4 @@
+export * from './Locked';
+export * from './Upload';
+export * from './Uploading';
+export * from './Uploaded';
